requirements.txt
web3==6.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
python-dotenv==1.0.0
requests==2.31.0
ccxt==3.1.54
joblib==1.3.1
tensorflow==2.13.0
xgboost==1.7.6
flask==2.3.3
flask-cors==4.0.0 